Add LifecycleType for KPACK
There are presently three LifecycleTypes supported: buildpack, docker and kpack. The first two have been supported, and this PR adds support for the third. It also includes a KpackData class for the lifecycle data that the kpack lifecycle type can return.

API documentation clarifies what is allowed:

http://v3-apidocs.cloudfoundry.org/version/3.122.0/index.html#lifecycles

Includes a test which expands and deserializes the Kpack LifecycleType.

Resolves #1094

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>